requirements.txt
pyqt5==5.15.9
keyboard==0.13.5
pyperclip==1.8.2
requests==2.31.0
playsound==1.3.0
pywin32==306
pillow==10.0.0
python-dotenv==1.0.0 